Refactoring: change package name
